Updated, improved version using yfinance
This is an upgraded version using yfinance instead of requests to scrape Yahoo Finance for data. Now also has the option to search for any stock index or other symbol available on Yahoo. Will prompt basic technical and fundamental data. Features an updated GUI, no longer bound to the terminal as in previous version.